readme added
    What is react and react Native?
    What is the difference between react and react native?
     
          History
    • https://en.wikipedia.org/wiki/React_(JavaScript_library)
    • https://en.wikipedia.org/wiki/React_Native

      What is React Native?
    • https://reactnative.dev/
    • https://www.oreilly.com/library/view/learning-react-native/9781491929049/ch01.html



      React vs React Native ….
    • https://www.javatpoint.com/reactjs-vs-reactnative
    • https://stackoverflow.com/questions/34641582/what-is-the-difference-between-react-native-and-react


	Youtube Short videos
    • https://www.youtube.com/watch?v=EVSMegdj6tY&t=493s
    • https://www.youtube.com/watch?v=6oFuwhIibo4


	Brief description of my Understandings : 
React is JS popular framework that is used for building responsive UI components. In react every thing is treated as a separate block like a building block. It is very responsive to changes like if there is any changes occure in DOM, it compare the actual DOM with virtual DOM and reflect the changes in actual DOM. So only that component is update where changes occures. 
React native is JS framework that is used for building native applications. “Learn once, Develop everywhere” 
This is framework is used to build application for cross platforms like IOS, windows, android mobile phones.